stats-cluster: add query_key to clusters
The goal is to provide an alternative stats query
to support a wider range of integration scenarios.

The center of this feature is the format of the query.
The query itself is a fully qualified name to a counter item,
or, a group of counters ('*' can be used in the query expression).

Examples:
center.queued.processed
center.*.processed
center.*.*

Query key

Current implementation does not change the way of how StatsClusters
are handled, just adds an extension to it: query_key is assigned to
each StatsCluster. This key is used as a lookup key when the query
expression is evaluated. Later implementations can build a stats-tree:
the keys are representing a path.

This patch initiates the query key concept.

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>